+++ src/modules/files/a.json
@@ -1,5 +1,5 @@
 {
-    "a": 1,
-    "b": 2,
-    "c": 3
+  "a": 1,
+  "b": 2,
+  "c": 3
 }
